Code cleaning dryness and improovements, Enhancing security of the application,
Previously

MODE_WORLD_READABLE:     (Creating world-readable files is very dangerous, and likely to cause security holes in applications. It is strongly discouraged; instead, applications should use more formal mechanism for interactions such as ContentProvider, BroadcastReceiver, and Service. There are no guarantees that this access mode will remain on a file, such as when it goes through a backup and restore. File creation mode: allow all other applications to have read access to the created file.)

was used which is deprecated and using that any other application can read the contents of the application is now enhanced.
and replaced with

MODE_PRIVATE: 	File creation mode: the default mode, where the created file can only be accessed by the calling application (or all applications sharing the same user ID.